+++ Create/Module2.bas
@@ -10,7 +10,7 @@ Attribute VB_Name = "Notes"
 
 
 ''===============================NOTE1 COMMON MESSAGE SENDER===============================
-'Dim showcnt As Integer, current As Integer
+'Dim showcnt As long, current As long
 'Sub NewMessage(Content As String, Color As Long, Optional ClearList As Boolean = False, Optional ClearOnly = False)
 '    current = -1
 '    If (ClearOnly And Not ClearList) Then
@@ -44,7 +44,7 @@ Attribute VB_Name = "Notes"
 '    Timer1.Interval = 1000
 'End Sub
 'Private Sub Timer1_Timer()
-'    Dim first As Integer
+'    Dim first As long
 '    If Timer1.Interval > 100 Then Timer1.Interval = Timer1.Interval - 100
 '    showcnt = showcnt + 1
 ''    If MsgContentList.ListCount <= 1 Then
@@ -62,6 +62,12 @@ Attribute VB_Name = "Notes"
 ''        If showcnt <> first Then ProgressBar.Width = showcnt / ShowCntPerMsg * Picture1.Width
 ''        Exit Sub
 ''    End If
+'    If current >= MsgContentList.ListCount Then
+'        Message.Caption = "No new messages."
+'        Message.ForeColor = vbWhite
+'        showcnt = ShowCntPerMsg - 1
+'        GoTo rrr
+'    End If
 '    If MsgContentList.ListCount = 0 Then
 '        Message.Caption = "No new messages."
 '        Message.ForeColor = vbWhite
